requirements.txt
Flask==1.0.2
google-cloud-firestore==1.6.0
